backend-azure/requirements.txt
azure-functions
numpy==1.24.4
Pillow
onnxruntime
azure-storage-blob
psycopg2-binary